high-requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
python-dotenv
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart